Uploading movies
These are the movie files that correspond to Fig. 4 of the paper.</br>
stiffness-total.gif - Fig. 4a -  Total shear stiffness of the network as a function of local shear (σ) and bend (ρ) moduli.</br>
stiffness-stretch.gif - Fig. 4b -  Stretch component of the total shear stiffness of the network as a function of local shear (σ) and bend (ρ) moduli.</br>
stiffness-bend.gif - Fig. 4c -  Bending component of the total shear stiffness of the network as a function of local shear (σ) and bend (ρ) moduli.</br>
stiffness-shear.gif - Fig. 4d -  Shear component of the total shear stiffness of the network as a function of local shear (σ) and bend (ρ) moduli.